1_SQL/requirements.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
PyMySQL==1.0.2
python-decouple==3.6
